+++ src/cli/test/Armadillo/Test/CliCommand.hs
@@ -8,6 +8,7 @@ module Armadillo.Test.CliCommand(
   RunningCliProcess(..),
   withCliCommand,
   runCliCommand,
+  CliLog(..),
 
   -- * Running the HTTP server
   RunningHttpServer(..),
@@ -16,12 +17,16 @@ module Armadillo.Test.CliCommand(
 ) where
 
 import qualified Armadillo.Api               as Api
-import           Armadillo.Cli.Command       (Command (..), ServerConfig (..))
+import           Armadillo.Cli.Command       (Command (..),
+                                              NodeClientConfig (..),
+                                              RefScriptCommand (..),
+                                              ServerConfig (..),
+                                              WalletClientOptions (..))
 import           Control.Concurrent          (threadDelay)
 import           Control.Monad               (void)
 import           Control.Tracer              (Tracer, traceWith)
-import           Convex.Devnet.Utils         (checkProcessHasNotDied, failure,
-                                              withLogFile)
+import           Convex.Devnet.Utils         (failure, withLogFile)
+import           Convex.Wallet.Operator      (OperatorConfigSigning (..))
 import           Data.Aeson                  (FromJSON, ToJSON)
 import           Data.Text                   (Text)
 import qualified Data.Text                   as Text
@@ -30,8 +35,7 @@ import           GHC.IO.Exception            (ExitCode (ExitSuccess))
 import           GHC.IO.Handle.Types         (Handle)
 import           Network.HTTP.Client         (defaultManagerSettings,
                                               newManager)
-import           Servant.Client              (ClientEnv, ClientError (..),
-                                              client, mkClientEnv, runClientM)
+import           Servant.Client              (ClientEnv, mkClientEnv)
 import           Servant.Client.Core.BaseUrl (BaseUrl (..), Scheme (..))
 import           System.FilePath             ((</>))
 import           System.IO                   (BufferMode (NoBuffering),
@@ -76,7 +80,8 @@ runCliCommand tracer stateDirectory command =
 commandString :: Command -> String
 commandString = \case
   StartServer{} -> "start-server"
-  WriteAPIFile{} -> "write-api-file"
+  WriteAPIFile{} -> "write-api"
+  RefScript{} -> "reference-scripts"
 
 cliProcess :: Maybe FilePath -> Command -> CreateProcess
 cliProcess cwd command = (proc cliExecutable strArgs){cwd} where
@@ -87,13 +92,57 @@ cliProcess cwd command = (proc cliExecutable strArgs){cwd} where
     WriteAPIFile{filePath} -> ["--api.file", filePath]
     _ -> []
 
+  refCommand = \case
+    RefScript _ c -> case c of
+      Deploy{} -> ["deploy"]
+      Check{}  -> ["check"]
+    _ -> []
+
+  nodeClient = \case
+    RefScript cfg _ -> nodeClientConfig cfg
+    _ -> []
+
+  walletClientOptions' = \case
+    RefScript _ (Deploy wco _ _) -> walletClientOptions wco
+    _ -> []
+
+  operatorSigningConfig' = \case
+    RefScript _ (Deploy _ osc _) -> operatorSigningConfig osc
+    _ -> []
+
+  outFile = \case
+    RefScript _ (Deploy _ _ f) -> ["--out.file", f]
+    _ -> []
+
   strArgs =
     mconcat
       [ [commandString command]
       , serverPort command
       , apiFile command
+      , nodeClient command
+      , refCommand command
+      , walletClientOptions' command
+      , operatorSigningConfig' command
+      , outFile command
       , ["+RTS", "-N2"]
       ]
+
+walletClientOptions :: WalletClientOptions -> [String]
+walletClientOptions WalletClientOptions{wcoHost, wcoPort} =
+  [ "--wallet.host", wcoHost
+  , "--wallet.port", show wcoPort
+  ]
+
+operatorSigningConfig :: OperatorConfigSigning -> [String]
+operatorSigningConfig OperatorConfigSigning{ocSigningKeyFile, ocStakeVerificationKeyFile} =
+  ["--signing-key-file", ocSigningKeyFile]
+  ++ maybe [] (\f -> ["--stake-verification-key-file", f]) ocStakeVerificationKeyFile
+
+nodeClientConfig :: NodeClientConfig -> [String]
+nodeClientConfig NodeClientConfig{nccCardanoNodeSocket, nccCardanoNodeConfigFile} =
+  [ "--node.socket", nccCardanoNodeSocket
+  , "--node.config", nccCardanoNodeConfigFile
+  ]
 
 cliExecutable :: String
 cliExecutable = "armadillo-cli"