+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Add CLI commands
   - `start-server`
   - `write-api-file`
+  - `reference-scripts`
+    - `deploy`
+    - `check`